Moving transaction related functions to their own dedicated file
